Automatic calculation of a and b coefficients
Update to automatically calculate the a and b coefficients (at the edges) for an arbitrary set of model levels.

Also added a preselection of reasonable model levels. For the current selection (22 levels) the average (absolute) error is about 1.6/1.7 % (a large part is also due to neglecting of different phases of water than the vapour phase).