Add config object with validation and docstrings (#165)
This PR adds a Config class to the model code. This loads the settings and also does some validation, e.g. whether a paths exist or a date is correctly formatted. Settings can now be accessed with dot notation instead of brackets.

In addition, there is now extensive documentation on the configuration options. This documentation is generated automatically based on the docstrings in config.py (using sphinx-autodoc)

The PR also removes the "/backtrack" in the path of the output folder (see #137), and moves the example config file(s).
Dates in existing config files should be reformatted (YYYY-MM-DD instead of YYYYMMDD).